do not show decimals for turns in character story file

+++ dump.go
@@ -98,7 +98,7 @@ func (g *game) Dump() string {
 	if g.Stats.Killed > 0 {
 		fmt.Fprintf(buf, "%d monsters died.\n", g.Stats.Killed)
 	}
-	fmt.Fprintf(buf, "You spent %.1f turns in Hareka's Underground.\n", float64(g.Turn)/10)
+	fmt.Fprintf(buf, "You spent %d turns in Hareka's Underground.\n", g.Turn/10)
 	maxDepth := Max(g.Depth, g.ExploredLevels)
 	s := "s"
 	if maxDepth == 1 {
@@ -172,9 +172,9 @@ func (g *game) DetailedStatistics(w io.Writer) {
 	if g.Stats.Digs > 0 {
 		fmt.Fprintf(w, "There were %d destroyed walls.\n", g.Stats.Digs)
 	}
-	fmt.Fprintf(w, "You spent %.1f%% turns wounded.\n", float64(g.Stats.TWounded)*100/float64(g.Stats.Turns+1))
-	fmt.Fprintf(w, "You spent %.1f%% turns with monsters in sight.\n", float64(g.Stats.TMonsLOS)*100/float64(g.Stats.Turns+1))
-	fmt.Fprintf(w, "You spent %.1f%% turns wounded with monsters in sight.\n", float64(g.Stats.TMWounded)*100/float64(g.Stats.Turns+1))
+	fmt.Fprintf(w, "You spent %d%% turns wounded.\n", g.Stats.TWounded*100/(g.Stats.Turns+1))
+	fmt.Fprintf(w, "You spent %d%% turns with monsters in sight.\n", g.Stats.TMonsLOS*100/(g.Stats.Turns+1))
+	fmt.Fprintf(w, "You spent %d%% turns wounded with monsters in sight.\n", g.Stats.TMWounded*100/(g.Stats.Turns+1))
 	maxDepth := Max(g.Depth-1, g.ExploredLevels)
 	if g.Player.HP <= 0 {
 		maxDepth++
@@ -362,7 +362,7 @@ func (g *game) SimplifedDump(err error) string {
 	} else {
 		fmt.Fprint(buf, "You did not recover the Gem Portal Artifact.\n")
 	}
-	fmt.Fprintf(buf, "You spent %.1f turns in Hareka's Underground.\n", float64(g.Turn)/10)
+	fmt.Fprintf(buf, "You spent %d turns in Hareka's Underground.\n", g.Turn/10)
 	maxDepth := Max(g.Depth, g.ExploredLevels)
 	s := "s"
 	if maxDepth == 1 {
